requirements.txt
Flask
Flasgger
python-dotenv
Flask-SQLAlchemy
Flask-Limiter
Flask-HTTPAuth
PyJWT
werkzeug